requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.4.1
    # via django
backcall==0.2.0
    # via ipython
brotli==1.0.9
    # via whitenoise
click==8.0.1
    # via pip-tools
decorator==5.0.9
    # via
    #   ipython
    #   traitlets
django==3.2.7
    # via
    #   -r requirements.in
    #   django-extensions
django-environ==0.6.0
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
factory-boy==3.2.0
    # via -r requirements.in
faker==8.12.1
    # via
    #   -r requirements.in
    #   factory-boy
honcho==1.0.1
    # via -r requirements.in
importlib-metadata==4.8.1
    # via
    #   click
    #   pep517
ipython==7.16.1
    # via -r requirements.in
ipython-genutils==0.2.0
    # via traitlets
jedi==0.17.2
    # via ipython
mock==4.0.3
    # via -r requirements.in
parso==0.7.1
    # via jedi
pep517==0.11.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.2.0
    # via -r requirements.in
prompt-toolkit==3.0.20
    # via ipython
psycopg2-binary==2.9.1
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pygments==2.10.0
    # via ipython
python-dateutil==2.8.2
    # via faker
pytz==2021.1
    # via django
six==1.16.0
    # via
    #   python-dateutil
    #   traitlets
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via faker
tomli==1.2.1
    # via pep517
traitlets==4.3.3
    # via ipython
typing-extensions==3.10.0.2
    # via
    #   asgiref
    #   importlib-metadata
waitress==2.0.0
    # via -r requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.0
    # via pip-tools
whitenoise[brotli]==5.3.0
    # via -r requirements.in
zipp==3.5.0
    # via
    #   importlib-metadata
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
